Add explicit typecast so that the result will be signed when calculating the offset to adjust when looking for the 'xref' marker.

+++ mupdf/pdf_open.c
@@ -113,7 +113,7 @@ pdf_readoldtrailer(pdf_xref *xref, char *buf, int cap)
 		/* broken pdfs where the section is not on a separate line */
 		if (s && *s != '\0')
 		{
-			error = fz_seek(xref->file, -(2 + strlen(s)), 1);
+			error = fz_seek(xref->file, -(2 + (int)strlen(s)), 1);
 			if (error)
 				return fz_rethrow(error, "cannot seek in file");
 		}
@@ -241,7 +241,7 @@ pdf_readoldxref(fz_obj **trailerp, pdf_xref *xref, char *buf, int cap)
 		if (s && *s != '\0')
 		{
 			fz_warn("broken xref section. proceeding anyway.");
-			error = fz_seek(xref->file, -(2 + strlen(s)), 1);
+			error = fz_seek(xref->file, -(2 + (int)strlen(s)), 1);
 			if (error)
 				return fz_rethrow(error, "cannot seek to xref");
 		}
